Don't append file extensions to request _by default_.
The HTTP specification defines two headers which should be used to
indicate the type of content being sent by a client, and the types of
content the client may accept in response; Content-Type and Accept.

Since some clients couldn't set these headers, some frameworks adopted a
workaround which was to permit specifying the content type as part of
the URL, for example "/books.xml", or "/books.json".

However, since setting headers is no problem for us, we don't do it, and
instead trust the upstream provider to correctly handle the Content-Type
and Accept headers (most web frameworks do this just fine).

An extra option is now supported by DataMapper.setup(...) which reverts
dm-rest-adapter to the previous behaviour:

    DataMapper.setup(:default, {
      :adapter  => 'rest',
      :format   => 'xml',
      # [...]
      :extension => true
    })

See http://jacobian.org/writing/rest-worst-practices/