Support validation errors from the API (via 422 response).
In the event that a create or update operation fails with a 422
Unprocessable Entity response, the adapter will take the errors returned
by the server, and add them as validation errors on the resource.

At present, however, the resource will still be marked as persisted /
clean since there doesn't appear to be a way to tell dm-core that the
operation failed. Needs further investigation...